Add analyzer as part of AspNet project
