+++ extensions/Worker.Extensions.Http.AspNetCore/src/Worker.Extensions.Http.AspNetCore.csproj
@@ -1,24 +1,53 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
-  <PropertyGroup>
-    <AssemblyName>Microsoft.Azure.Functions.Worker.Extensions.Http.AspNetCore</AssemblyName>
-    <RootNamespace>Microsoft.Azure.Functions.Worker.Extensions.Http.AspNetCore</RootNamespace>
-    <Description>ASP.NET Core extensions for .NET isolated functions</Description>
+    <PropertyGroup>
+        <AssemblyName>Microsoft.Azure.Functions.Worker.Extensions.Http.AspNetCore</AssemblyName>
+        <RootNamespace>Microsoft.Azure.Functions.Worker.Extensions.Http.AspNetCore</RootNamespace>
+        <Description>ASP.NET Core extensions for .NET isolated functions</Description>
+        <!--Version information-->
+        <VersionPrefix>1.44.0</VersionPrefix>
+        <TargetFrameworks>net6.0; netstandard2.0</TargetFrameworks>
+    </PropertyGroup>
 
-    <!--Version information-->
-    <VersionPrefix>1.0.0</VersionPrefix>
-    <TargetFramework>net6.0</TargetFramework>
-  </PropertyGroup>
+    <PropertyGroup Condition=" '$(TargetFramework)' == 'netstandard2.0'">
+        <OutputType>Library</OutputType>
+        <Nullable>disable</Nullable>
+        <SuppressDependenciesWhenPacking>true</SuppressDependenciesWhenPacking>
+        <IncludeBuildOutput>false</IncludeBuildOutput>
+        <PackageTags>Azure Functions Worker, analyzers</PackageTags>
+        <TargetsForTfmSpecificContentInPackage>$(TargetsForTfmSpecificContentInPackage);_AddAnalyzersToOutput</TargetsForTfmSpecificContentInPackage>
+    </PropertyGroup>
 
-  <Import Project="..\..\..\build\Extensions.props" />
-  <ItemGroup>
-    <FrameworkReference Include="Microsoft.AspNetCore.App" />
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\..\..\src\DotNetWorker\DotNetWorker.csproj" />
-    <ProjectReference Include="..\..\Worker.Extensions.Abstractions\src\Worker.Extensions.Abstractions.csproj" />
-    <ProjectReference Include="..\..\Worker.Extensions.Http\src\Worker.Extensions.Http.csproj" />
-    <ProjectReference Include="..\..\..\sdk\Sdk.Analyzers\Sdk.Analyzers.csproj" />
-  </ItemGroup>
+
+    <Import Project="..\..\..\build\Extensions.props" />
+
+    <ItemGroup Condition=" '$(TargetFramework)' == 'net6.0'">
+        <FrameworkReference Include="Microsoft.AspNetCore.App" />
+    </ItemGroup>
+    
+    <ItemGroup Condition=" '$(TargetFramework)' == 'net6.0'">
+        <ProjectReference Include="..\..\..\src\DotNetWorker\DotNetWorker.csproj" />
+        <ProjectReference Include="..\..\Worker.Extensions.Abstractions\src\Worker.Extensions.Abstractions.csproj" />
+        <ProjectReference Include="..\..\Worker.Extensions.Http\src\Worker.Extensions.Http.csproj" />
+    </ItemGroup>
+    
+    <ItemGroup Condition=" '$(TargetFramework)' == 'netstandard2.0'">
+        <Compile Remove="**\*.cs; *.cs" />
+        <Compile Include="Analyzer\*.cs" />
+    </ItemGroup>
+
+    <ItemGroup>
+        <PackageReference Include="Microsoft.CodeAnalysis.Analyzers" Version="3.3.3">
+            <PrivateAssets>all</PrivateAssets>
+            <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+        </PackageReference>
+        <PackageReference Include="Microsoft.CodeAnalysis.CSharp.Workspaces" Version="3.11.0" />
+    </ItemGroup>
+
+    <Target Name="_AddAnalyzersToOutput">
+        <ItemGroup>
+            <TfmSpecificPackageFile Include="$(OutputPath)\Microsoft.Azure.Functions.Worker.Extensions.Http.AspNetCore.dll" PackagePath="analyzers/dotnet/cs" />
+        </ItemGroup>
+    </Target>
 
 </Project>
